feat: fix cast long to biginteger for report context; change version to 5.1.22-SNAPSHOT;

+++ src/main/java/org/qubership/automation/itf/core/hibernate/spring/managers/reports/InstanceContextObjectManager.java
@@ -141,7 +141,7 @@ public class InstanceContextObjectManager extends AbstractObjectManager<Instance
         if (!(Objects.isNull(initiatorInfo) || initiatorInfo.isEmpty())) {
             tcContextBriefInfo.setIniname((String) initiatorInfo.get(0)[0]);
             tcContextBriefInfo.setInitiatortype((String) initiatorInfo.get(0)[1]);
-            tcContextBriefInfo.setSituationId((BigInteger) initiatorInfo.get(0)[2]);
+            tcContextBriefInfo.setSituationId(BigInteger.valueOf((Long) initiatorInfo.get(0)[2]));
             tcContextBriefInfo.setChainId(BigInteger.valueOf((Long) initiatorInfo.get(0)[3]));
             tcContextBriefInfo.setExecutiondata((String) initiatorInfo.get(0)[4]);
             tcContextBriefInfo.setOperationName((String) initiatorInfo.get(0)[5]);
